requirements.txt
asgiref==3.8.1
Django==3.2.25
django-cors-headers==4.5.0
django-extensions==3.2.3
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
pillow==11.2.1
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-decouple==3.8
python-dotenv==1.1.0
pytz==2025.2
sqlparse==0.5.3
typing_extensions==4.13.2
pytest
pytest-django
drf-spectacular


